# docs/swagger/loyaltysystem-api-docs.yaml
openapi: 3.0.3
info:
  title: Loyalty System API
  description: |
    This is the API documentation for the Loyalty System platform, covering all the endpoints required for:
    * Customer App - Mobile application for end customers
    * Staff App - Application for store staff to handle in-store operations
    * Admin App - Administrative portal for business owners and managers
  version: 1.0.0
servers:
  - url: https://api.loyaltysystem.com/v1
    description: Production server
  - url: https://staging-api.loyaltysystem.com/v1
    description: Staging server
  - url: http://localhost:5000
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoyaltyProgram:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        businessId:
          type: string
          format: uuid
        brandId:
          type: string
          format: uuid
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        rewards:
          type: array
          items:
            $ref: '#/components/schemas/Reward'
    LoyaltyProgramCreateRequest:
      type: object
      required:
        - name
        - description
        - businessId
        - brandId
      properties:
        name:
          type: string
        description:
          type: string
        businessId:
          type: string
          format: uuid
        brandId:
          type: string
          format: uuid
        isActive:
          type: boolean
          default: true

    # Loyalty Card schemas
...
    Reward:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        loyaltyProgramId:
          type: string
          format: uuid
        pointsCost:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
...
    PaginatedResponse:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
        totalItems:
          type: integer
        pageNumber:
          type: integer
        pageSize:
          type: integer
        totalPages:
          type: integer

    # Error response
...
security:
  - bearerAuth: []
paths:
...
  /api/loyaltyPrograms/{id}:
    get:
      tags:
        - Loyalty Programs
      summary: Get loyalty program by ID
      description: Returns details of a specific loyalty program
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved loyalty program
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoyaltyProgram'
        '404':
          description: Program not found
        '401':
          description: Unauthorized
...
  /api/loyaltyPrograms/{id}/rewards:
    get:
      tags:
        - Rewards
      summary: Get rewards for a loyalty program
      description: Returns all active rewards available for a specific loyalty program
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved rewards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Reward'
        '404':
          description: Program not found
        '401':
          description: Unauthorized
  /api/loyaltyPrograms/{id}/rewards/{rewardId}:
    get:
      tags:
        - Rewards
      summary: Get specific reward for a loyalty program
      description: Returns details of a specific reward for a loyalty program
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: rewardId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved reward
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reward'
        '404':
          description: Program or reward not found
        '401':
          description: Unauthorized
  /api/loyaltyPrograms/nearby:
    get:
      tags:
        - Loyalty Programs
      summary: Get nearby loyalty programs
      description: Returns loyalty programs near a geographic location
      parameters:
        - name: latitude
          in: query
          required: true
          schema:
            type: number
            format: double
        - name: longitude
          in: query
          required: true
          schema:
            type: number
            format: double
        - name: radiusKm
          in: query
          schema:
            type: number
            format: double
            default: 10
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Successfully retrieved nearby programs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
        '400':
          description: Invalid parameters
        '401':
          description: Unauthorized
  /api/loyaltyPrograms/search:
    get:
      tags:
        - Loyalty Programs
      summary: Search loyalty programs
      description: Searches for loyalty programs matching a query
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Successfully retrieved search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'
        '400':
          description: Invalid parameters
        '401':
          description: Unauthorized

  # Loyalty Cards (Customer)
...
  /api/admin/loyalty-programs/{id}:
    get:
      tags:
        - Loyalty Programs
      summary: Get loyalty program by ID
      description: Returns details of a specific loyalty program
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved loyalty program
        '404':
          description: Program not found
        '401':
          description: Unauthorized
    put:
      tags:
        - Loyalty Programs
      summary: Update a loyalty program
      description: Updates details of a specific loyalty program
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoyaltyProgramCreateRequest'
      responses:
        '200':
          description: Successfully updated loyalty program
        '404':
          description: Program not found
        '400':
          description: Bad request
        '401':
          description: Unauthorized
    delete:
      tags:
        - Loyalty Programs
      summary: Delete a loyalty program
      description: Deletes a specific loyalty program
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Successfully deleted loyalty program
        '404':
          description: Program not found
        '401':
          description: Unauthorized

  # Rewards (Admin)
...